% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.zero.r
\name{n.zero}
\alias{n.zero}
\title{Compter les zéros}
\usage{
n.zero(x)
}
\arguments{
\item{x}{un vecteur de type 'integer'}
}
\value{
le nombre de 0 dans x
}
\description{
Compter les zéros
}
\details{
Cette fonction démontre l'utilisation d'une boucle C++11
pour compter les zéros dans un vecteur.
}
\examples{
a <- sample(0:99, 1e6, TRUE )
n.zero(a)

}
