% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.r
\name{split}
\alias{split}
\title{Fonction split}
\usage{
split(x, pivot)
}
\arguments{
\item{x}{un vecteur d'entiers ou de doubles}

\item{pivot}{(facultatif) un pivot}
}
\value{
une liste avec composantes `lo` et `hi`
}
\description{
Fonction split
}
\details{
si `pivot` est absent, 
un élement de `x` pris au hasard sera utilisé
}
\examples{
x <- runif(20)
split(x, 0.5)

x <- sample.int(100, 20)
split(x, 50)
split(x)
}
